Add files via upload
results_database.sqlite - results from FERROSCOPE on all polar structures in the CSD (as of Sep24 release). RMS distribution given in Figure 8b of the paper.
results_database_nonpolar.sqlite - results from FERROSCOPE on all non-polar structures in the CSD (as of Sep24 release). RMS distribution given in Figure 11d of the paper.